% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_accessibility.R
\name{accessible_action_link}
\alias{accessible_action_link}
\title{accessible_action_link}
\usage{
accessible_action_link(inputId, label, icon = NULL, ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{icon}{An optional \code{\link[shiny:icon]{icon()}} to appear on the button.}

\item{...}{Named attributes to be applied to the button or link.}
}
\value{
HTML
}
\description{
accessible_action_link
}
\examples{
accessible_action_link("infoLink", "Information Link", class = "btn-info")
}
