% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faithful.R
\docType{data}
\name{faithful}
\alias{faithful}
\title{Old Faithful Geyser Data}
\format{
A data frame with 272 observations on 2 variables.
}
\source{
{W. H<U+00E4>rdle.}
}
\usage{
faithful
}
\description{
A closer look at \code{faithful$eruptions} reveals that these are
heavily rounded times originally in seconds, where multiples of 5 are
more frequent than expected under non-human measurement.  For a
better version of the eruption times, see the example below.
There are many versions of this dataset around: Azzalini and Bowman
(1990) use a more complete version.
}
\details{
\itemize{
    \item eruptions. Eruption time in mins
    \item waiting. Waiting time to next eruption (in mins)
}
}
\keyword{datasets}
