% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_review.R
\name{gen_template_doc}
\alias{gen_template_doc}
\title{Generate a initial Word doc to write draft text in}
\usage{
gen_template_doc(
  rv_dir = "inst/review",
  docx_file = "review.docx",
  styles_rmd = "inst/review/styles/draft-styles.rmd"
)
}
\arguments{
\item{rv_dir}{Review directory}

\item{docx_file}{Output Word document file name}

\item{styles_rmd}{Path to Rmarkdown that creates Word template}
}
\value{
Path to the generated Word document
}
\description{
The `{officer}` package works best with a Word document generated from
Rmarkdown. This is just a wrapper around `rmarkdown::render` to do that, with
default arguments matching our package conventions.
}
\examples{
\dontrun{
# For standard use case, default args are fine
gen_template_doc()

# May be times when you want to use outside of the usual use case, e.g.
# generate Word doc for adhoc purposes
gen_template_doc(
  "C:/Users/CYPHER/Downloads",
  "adhoc.docx",
  system.file(
    "inst", "review", "styles", "draft-styles.rmd", package = "nhsbsaShinyR"
  )
)}
}
