% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_accessibility.R
\name{h4_tabstop}
\alias{h4_tabstop}
\title{h4_tabstop}
\usage{
h4_tabstop(header, tabindex = 0, ...)
}
\arguments{
\item{header}{Heading text}

\item{tabindex}{Number for tabindex, default 0}

\item{...}{
  Arguments passed on to \code{\link[shiny:reexports]{shiny::h4}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
HTML
}
\description{
h4_tabstop
}
\examples{
h4_tabstop("Heading")
}
