% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_accessibility.R
\name{h5_tabstop}
\alias{h5_tabstop}
\title{h5_tabstop}
\usage{
h5_tabstop(heading, ..., tabindex = 0)
}
\arguments{
\item{heading}{Heading text}

\item{...}{
  Arguments passed on to \code{\link[shiny:reexports]{shiny::h1}}
  \describe{
    \item{\code{}}{}
  }}

\item{tabindex}{Number for tabindex, default 0}
}
\value{
HTML
}
\description{
h5_tabstop
}
\examples{
h5_tabstop("Heading")
}
