% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_accessibility.R
\name{h6_tabstop}
\alias{h6_tabstop}
\title{h6_tabstop}
\usage{
h6_tabstop(header, tabindex = 0, ...)
}
\arguments{
\item{header}{Heading text}

\item{tabindex}{Number for tabindex, default 0}

\item{...}{
  Arguments passed on to \code{\link[shiny:reexports]{shiny::h6}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
HTML
}
\description{
h6_tabstop
}
\examples{
h6_tabstop("Heading")
}
