% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_review.R
\name{md_to_word}
\alias{md_to_word}
\title{Generate Word doc from markdown files}
\usage{
md_to_word(
  md_dir = "inst/app/www/assets/markdown",
  rv_dir = "inst/review",
  docx_file = "review.docx",
  styles_rmd = "inst/review/styles/draft-styles.rmd"
)
}
\arguments{
\item{md_dir}{Markdown directory}

\item{rv_dir}{Review directory}

\item{docx_file}{Output Word document file name}

\item{styles_rmd}{Path to Rmarkdown that creates Word template}
}
\value{
Path of generated Word doc
}
\description{
All markdown files in a folder can be used to generate a Word document. The
main use case is for use in reviewing often changing text when writing final
output of initiatives. It can be used standalone for adhoc purposes.
}
\examples{
\dontrun{
# For standard use case, default args are fine
md_to_word()

# May be times when you want to use outside of the usual use case, e.g.
# generate Word doc for adhoc purposes
md_to_word(
  "my/adhoc/markdown",
  "C:/Users/CYPHER/Downloads",
  "adhoc.docx",
  system.file(
    "inst", "review", "styles", "draft-styles.rmd", package = "nhsbsaShinyR"
  )
)}
}
