% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_accessibility.R
\name{nhs_card_tabstop}
\alias{nhs_card_tabstop}
\title{nhs_card_tabstop}
\usage{
nhs_card_tabstop(heading, ..., tabindex = 0)
}
\arguments{
\item{heading}{Card title}

\item{...}{Card content}

\item{tabindex}{Number for tabindex, default 0}
}
\value{
HTML
}
\description{
nhs_card_tabstop
}
\examples{
nhs_card_tabstop("A card", shiny::p("Some content"))
}
