% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_footer.R
\name{nhs_footer}
\alias{nhs_footer}
\title{nhs_footer}
\usage{
nhs_footer(
  a11y_statement_url = "#",
  email = "nhsbsa.dall@nhs.net",
  github_url = "https://github.com/nhsbsa-data-analytics"
)
}
\arguments{
\item{a11y_statement_url}{URL of accessibility statement}

\item{email}{Contact email, by default 'nhsbsa.dall@nhs.net'}

\item{github_url}{URL of app GitHub repo, by default 
https://github.com/nhsbsa-data-analytics}
}
\value{
HTML
}
\description{
NHS style footer element
}
\examples{
nhs_footer()
}
