% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_grid.R
\name{nhs_grid_2_col}
\alias{nhs_grid_2_col}
\title{nhs_grid_2_col}
\usage{
nhs_grid_2_col(col_1, col_2)
}
\arguments{
\item{col_1}{First column content}

\item{col_2}{Second column content}
}
\value{
HTML
}
\description{
NHS style 2 column grid element
}
\examples{
nhs_grid_2_col(shiny::p("Column one"), shiny::p("Column two"))
}
