% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_grid.R
\name{nhs_grid_4_col}
\alias{nhs_grid_4_col}
\title{nhs_grid_4_col}
\usage{
nhs_grid_4_col(col_1, col_2, col_3, col_4)
}
\arguments{
\item{col_1}{First column content}

\item{col_2}{Second column content}

\item{col_3}{Third column content}

\item{col_4}{Fourth column content}
}
\value{
HTML
}
\description{
NHS style 4 column grid element
}
\examples{
nhs_grid_4_col(
  shiny::p("Column one"),
  shiny::p("Column two"),
  shiny::p("Column three"),
  shiny::p("Column four")
)
}
