% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_navlistPanel.R
\name{nhs_navlistPanel}
\alias{nhs_navlistPanel}
\title{nhs_navlistPanel}
\usage{
nhs_navlistPanel(
  ...,
  id = NULL,
  selected = NULL,
  header = NULL,
  footer = NULL,
  fluid = TRUE,
  widths = c(4, 8)
)
}
\arguments{
\item{...}{\code{\link[shiny:tabPanel]{tabPanel()}} elements to include in the navlist}

\item{id}{If provided, you can use \verb{input$}\emph{\code{id}} in your
server logic to determine which of the current navlist items is active. The
value will correspond to the \code{value} argument that is passed to
\code{\link[shiny:tabPanel]{tabPanel()}}.}

\item{selected}{The \code{value} (or, if none was supplied, the \code{title})
of the navigation item that should be selected by default. If \code{NULL},
the first navigation will be selected.}

\item{header}{Tag or list of tags to display as a common header above all
tabPanels.}

\item{footer}{Tag or list of tags to display as a common footer below all
tabPanels}

\item{fluid}{\code{TRUE} to use fluid layout; \code{FALSE} to use fixed
layout.}

\item{widths}{Column widths of the navigation list and tabset content areas
respectively.}
}
\value{
HTML
}
\description{
NHS style navlist element
}
\examples{
nhs_navlistPanel(
  "nav_panel",
  shiny::tabPanel(title = "Introduction", shiny::p("This is an intro")),
  shiny::tabPanel(title = "Article", shiny::p("This is an article"))
)
}
