% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhsbsaShinyR.R
\docType{package}
\name{nhsbsaShinyR-package}
\alias{nhsbsaShinyR}
\alias{nhsbsaShinyR-package}
\title{nhsbsaShinyR: Template for NHSBSA {shiny} apps}
\description{
Template used for NHSBSA Data Science and Advanced Analytics {shiny} apps.
}
\author{
\strong{Maintainer}: DALL \email{nhsbsa.dall@nhs.net}

Other contributors:
\itemize{
  \item Adam Ivison [contributor]
  \item Kayoung Goffe \email{kayoung.goffe@nhsbsa.nhs.uk} [contributor]
  \item Steven Buckley \email{steven.buckley@nhsbsa.nhs.uk} [contributor]
  \item Mark McPherson \email{mark.mcphersoni@nhsbsa.nhs.uk} [contributor]
}

}
\keyword{internal}
