% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_review.R
\name{review_md}
\alias{review_md}
\title{Compare and create snapshot for a single markdown file}
\usage{
review_md(path, snaps_dir = "inst/review/tests")
}
\arguments{
\item{path}{The path of the markdown file}

\item{snaps_dir}{The path of the directory containing (or to create in) the
_snaps directory}
}
\value{
Nothing, used for side-effects only
}
\description{
Will announce intention to compare/create snapshots for each markdown file,
then do so.
}
\examples{
\dontrun{
review_md(
  path = "my/adhoc/markdown/md_file.md",
  snaps_dir = "C:/Users/CYPHER/Downloads"
)
)}
}
