% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_review.R
\name{review_md_diff}
\alias{review_md_diff}
\title{Review changes to markdown files}
\usage{
review_md_diff(snap_dir = "review_md/", snaps_dir = "inst/review/tests")
}
\arguments{
\item{snap_dir}{The name of the directory the snapshots are in}

\item{snaps_dir}{The path of the directory containing (or to create in) the
_snaps directory}
}
\value{
Nothing, used for side-effects only
}
\description{
This is a customised combination of `testthat::snapshot_review` and the
unexported function `testthat:::review_app`. The purpose is to allow user to
see the changes made to markdown for the app, and then approve or reject
them. This is repeated, with changes to the markdown, until all changes are
accepted.
}
\examples{
\dontrun{
# For standard use case, default args are fine
review_md_diff()

# Can be paired with review_md_dir() for reviewing custom markdown
review_md_dir(
  md_dir = "my/adhoc/markdown",
  snaps_dir = "C:/Users/CYPHER/Downloads"
)

review_md_diff(snaps_dir = "C:/Users/CYPHER/Downloads")
)}
}
