% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_review.R
\name{review_md_dir}
\alias{review_md_dir}
\title{Compare and create snapshots for all markdown files in a directory}
\usage{
review_md_dir(md_dir = "inst/review/temp", snaps_dir = "inst/review/tests")
}
\arguments{
\item{md_dir}{The path to the directory of markdown files}

\item{snaps_dir}{The path of the directory containing (or to create in) the
_snaps directory}
}
\value{
Nothing, used for side-effects only
}
\description{
Will announce intention to compare/create snapshots for each markdown file,
then do so.
}
\examples{
\dontrun{
# For standard use case, default args are fine
review_md_dir()

# Can be paired with review_md_diff() for reviewing custom markdown
review_md_dir(
  md_dir = "my/adhoc/markdown",
  snaps_dir = "C:/Users/CYPHER/Downloads"
)

review_md_diff(snaps_dir = "C:/Users/CYPHER/Downloads")
)}
}
