% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{theme_nhsbsa}
\alias{theme_nhsbsa}
\title{Custom NHSBSA highcharter theme}
\usage{
theme_nhsbsa(hc, palette = NA, stack = "normal")
}
\arguments{
\item{palette}{Which colour palette to use from the `nhsbsaR` package.}

\item{stack}{Stack option for highcharter.}
}
\value{

}
\description{
Based on the nhsbsaR highcharter theme, since it returns a list we can edit
it to the specific theme for this shiny app.
}
