% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_review.R
\name{update_md}
\alias{update_md}
\title{Move all markdown files from one directory to another}
\usage{
update_md(
  md_tmp_dir = "inst/review/temp",
  md_dir = "inst/app/www/assets/markdown"
)
}
\arguments{
\item{md_tmp_dir}{The path to the directory of reviewed markdown files}

\item{md_dir}{The path to the directory for markdown files used by app}
}
\value{
Nothing, used for side-effects only
}
\description{
Convenience function to quickly move all the reviewed new markdown files to
the app markdown folder.
}
\examples{
\dontrun{
# For standard use case, default args are fine
update_md()

# Unlikely needed, but can be used to move markdown files between any two folders
update_md(
  md_dir = "my/adhoc/markdown",
  snaps_dir = "my/final/markdown"
)}
}
