% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_review.R
\name{word_to_md}
\alias{word_to_md}
\title{Generate markdown files from Word doc}
\usage{
word_to_md(
  md_flag = "markdown/",
  rv_dir = "inst/review",
  docx_file = "review.docx",
  md_out_dir = "inst/review/temp",
  first_run_snaps = TRUE
)
}
\arguments{
\item{md_flag}{How markdown file is flagged in Word doc; lines starting with
this will be taken as flags to assign following content to a markdown file
with name as everything after this}

\item{rv_dir}{Review directory}

\item{docx_file}{Output Word document file name}

\item{md_out_dir}{Output folder for markdown files}

\item{first_run_snaps}{Should snapshot files be generated if there are none?}
}
\value{
Nothing, used for side effects only
}
\description{
A Word document can be used to generate multiple markdown files. The main use
case is for use in reviewing often changing text when writing final output of
initiatives. It can be used standalone for adhoc purposes.
}
\examples{
\dontrun{
# For standard use case, default args are fine
word_to_md()

# May be times when you want to use outside of the usual use case, e.g.
# generate markdown for adhoc purposes
word_to_md(
  "my/adhoc/markdown",
  "C:/Users/CYPHER/Downloads",
  "adhoc.docx",
  "C:/Users/CYPHER/Downloads"
)}
}
