% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhs_selectInput.R
\name{nhs_selectInput}
\alias{nhs_selectInput}
\title{nhs_selectInput Function}
\usage{
nhs_selectInput(inputId, label, choices, selected = NULL, full_width = FALSE)
}
\description{
nhs_selectInput Function
}
